// packages/dragon/src/forms/ua-shadow.generated.ts
export const RANGE_UA_SHADOW = {
  "container": [
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "display",
          "block",
          false
        ]
      ]
    },
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "unicode-bidi",
          "isolate",
          false
        ]
      ]
    },
    {
      "selectors": [
        "input[type=\"range\" i]::-webkit-slider-container"
      ],
      "declarations": [
        [
          "appearance",
          "inherit",
          false
        ],
        [
          "flex-grow",
          "1",
          false
        ],
        [
          "flex-shrink",
          "1",
          false
        ],
        [
          "flex-basis",
          "0%",
          false
        ],
        [
          "min-inline-size",
          "0px",
          false
        ],
        [
          "box-sizing",
          "border-box",
          false
        ],
        [
          "display",
          "flex",
          false
        ],
        [
          "-webkit-user-modify",
          "read-only !important",
          true
        ]
      ]
    }
  ],
  "track": [
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "display",
          "block",
          false
        ]
      ]
    },
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "unicode-bidi",
          "isolate",
          false
        ]
      ]
    },
    {
      "selectors": [
        "input[type=\"range\" i]::-webkit-slider-runnable-track"
      ],
      "declarations": [
        [
          "flex-grow",
          "1",
          false
        ],
        [
          "flex-shrink",
          "1",
          false
        ],
        [
          "flex-basis",
          "0%",
          false
        ],
        [
          "min-inline-size",
          "0px",
          false
        ],
        [
          "align-self",
          "center",
          false
        ],
        [
          "box-sizing",
          "border-box",
          false
        ],
        [
          "display",
          "block",
          false
        ],
        [
          "-webkit-user-modify",
          "read-only !important",
          true
        ]
      ]
    }
  ],
  "thumb": [
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "display",
          "block",
          false
        ]
      ]
    },
    {
      "selectors": [
        "div"
      ],
      "declarations": [
        [
          "unicode-bidi",
          "isolate",
          false
        ]
      ]
    },
    {
      "selectors": [
        "input[type=\"range\" i]::-webkit-slider-thumb"
      ],
      "declarations": [
        [
          "appearance",
          "auto",
          false
        ],
        [
          "box-sizing",
          "border-box",
          false
        ],
        [
          "display",
          "block",
          false
        ],
        [
          "-webkit-user-modify",
          "read-only !important",
          true
        ]
      ]
    }
  ]
} as const;

export const RANGE_THEME_THUMB = {"width":16,"height":16} as const;

export const RANGE_NONE_THUMB = {"width":129,"height":0} as const;